Fixing error of type casting (vector to Integer)
The other virtual machines send to physical machine a byte stream
containing value in type integer. But when the physical machine receive
that stream, it realizes that it has received a stream of type "Vector".
So I guess that there is an error while transferring data in system.
Even though, between 2 virtual machines, this casting error does not
happen, and the program work well. Only when the physical machine run,
this error is caused.