function : Adding File Torrent
This function is to broadcast the torrent of the file you want to
download to all other machines in network. It will compare this torrent
content with available torrent content in each machine. If they matches,
the machine contains the file you want to download. This machine will
reply to you and be ready to send you the chunks.